refactor(tool): vendor taint model files
The taint models from upstream pysa are somewhat outdated.
This commit adds the django ones and applies them conditionally,
depending of django is a dependency.
It also considers the HttpRequest object to be a source, which fixes
some cases where due lack of typing, pysa would not correctly taint
accesses to `request.body`.